[CLEANUP] Rename the "Mail" namespace to "Email"
Closes #478
