+++ Classes/Email/RealMailer.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OliverKlee\Oelib\Mail;
+namespace OliverKlee\Oelib\Email;
 
 use TYPO3\CMS\Core\Mail\MailMessage;
