+++ Tests/Unit/Email/RealMailerTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace OliverKlee\Oelib\Tests\Unit\Email;
 
 use Nimut\TestingFramework\TestCase\UnitTestCase;
-use OliverKlee\Oelib\Mail\Mail;
-use OliverKlee\Oelib\Mail\RealMailer;
+use OliverKlee\Oelib\Email\Mail;
+use OliverKlee\Oelib\Email\RealMailer;
 use OliverKlee\Oelib\Tests\Unit\Email\Fixtures\TestingMailRole;
 use PHPUnit\Framework\MockObject\MockObject;
 use TYPO3\CMS\Core\Mail\MailMessage;
